+++ src/requestsDebugger.js
@@ -11,7 +11,8 @@ var RdGlobalConfig = constants.RdGlobalConfig;
 var STATIC_MESSAGES = constants.STATIC_MESSAGES;
 var CommandLineManager = require('./commandLine');
 var ConnectivityChecker = require('./connectivity');
-var server = require('./server');
+var proxy = require('./proxy');
+var reverseProxy = require('./reverseProxy');
 var StatsFactory = require('./stats/statsFactory');
 var LogManager = require('./logger');
 var fs = require('fs');
@@ -139,7 +140,7 @@ var RdTool = {
       console.log(Utils.formatAndBeautifyLine(STATIC_MESSAGES.CONNECTIVITY_CHECKS_DONE, '', '-', 60, true));
     });
 
-    server.RdHandler.startProxy(RdGlobalConfig.RD_HANDLER_PORT, function (err, result) {
+    proxy.RdHandler.startServer(RdGlobalConfig.RD_HANDLER_PORT, function (err, result) {
       if (err) {
         console.log(STATIC_MESSAGES.ERR_STARTING_TOOL, err);
         console.log('Exiting the Proxy...');
@@ -148,7 +149,7 @@ var RdTool = {
       console.log(Utils.formatAndBeautifyLine(STATIC_MESSAGES.TOOL_STARTED_ON_PORT + result, '', '-', 60, true));
     });
 
-    server.RdReverseProxyHandler.startServer(RdGlobalConfig.RD_HANDLER_REVERSE_PROXY_PORT, function (err, result) {
+    reverseProxy.RdHandler.startServer(RdGlobalConfig.RD_HANDLER_REVERSE_PROXY_PORT, function (err, result) {
       if (err) {
         console.log(STATIC_MESSAGES.ERR_STARTING_TOOL, err);
         console.log('Exiting the Reverse Proxy...');
